reducer modularized
