+++ FRONT/src/Redux/reducer/ingredients.js
@@ -1,0 +1,17 @@
+import { GET_INGREDIENTS } from "../actions/index.js";
+
+const initialState = {
+  ingredients: null,
+};
+
+const ingredientsReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case GET_INGREDIENTS:
+      return { ...state, ingredients: action.payload };
+
+    default:
+      return { ...state };
+  }
+};
+
+export default ingredientsReducer;
